requirements/base.txt
pytz==2022.2.1  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.5  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower
sorl-thumbnail==12.9.0  # https://github.com/jazzband/sorl-thumbnail
sorl-thumbnail-serializer-field==0.2.1  # https://github.com/dessibelle/sorl-thumbnail-serializer-field
django-ordered-model==3.6  # https://github.com/django-ordered-model/django-ordered-model
weasyprint==56.1

# Django
# ------------------------------------------------------------------------------
django==4.1.9  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.0.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.26.2  # https://github.com/tfranzel/drf-spectacular
djangorestframework-simplejwt==5.2.0
django-filter==22.1
dj-rest-auth==2.2.5
djangorestframework-recursive==0.1.2
django-phone-auth==0.3.1
django-mptt
django_seed==0.3.1
django_ckeditor==6.5.1  # https://github.com/django-ckeditor/django-ckeditor
django-modeltranslation==0.18.5  # https://django-modeltranslation.readthedocs.io/en/latest/
# django-multiselectfield==0.1.12  # https://github.com/goinnn/django-multiselectfield
django-multiselectfield2==0.0.1  # https://github.com/goinnn/django-multiselectfield
django-jazzmin==2.6.0
